flask_AI2/requirements.txt
flask==2.0.3
ipython
joblib
pandas
parso
numpy==1.26.4
scikit-learn==1.2.2
urllib3
Werkzeug==2.0.3
python-telegram-bot
tensorflow
scipy